implemented method getAllByDriver in CarDaoJdbcImpl
