+++ src/main/java/core/basesyntax/dao/CarDaoJdbcImpl.java
@@ -105,21 +105,54 @@ public class CarDaoJdbcImpl implements CarDao {
 
     @Override
     public List<Car> getAllByDriver(Long driverId) {
-        return null;
+        String getAllByDriverQuery = "SELECT * FROM taxi_service.cars_drivers cd"
+                + " INNER JOIN taxi_service.cars c ON cd.car_id = c.id"
+                + " INNER JOIN taxi_service.drivers d ON cd.driver_id = d.id"
+                + " WHERE cd.driver_id = ?";
+        List<Car> cars = new ArrayList<>();
+        try (Connection connection = ConnectionUtil.getConnection();
+                 PreparedStatement preparedStatement = connection
+                         .prepareStatement(getAllByDriverQuery)) {
+            preparedStatement.setLong(1, driverId);
+            ResultSet resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                cars.add(getCar(resultSet));
+            }
+            return cars;
+        } catch (SQLException e) {
+            throw new DataProcessingException("Cannot get any car by driver by id " + driverId, e);
+        }
     }
 
     private Car getCar(ResultSet resultSet) throws SQLException {
         Long id = resultSet.getObject("c.id", Long.class);
         Long manufacturerId = resultSet.getObject("c.manufacturer_id", Long.class);
         String model = resultSet.getObject("c.model", String.class);
-        String nameManufacturer = resultSet.getObject("m.name", String.class);
-        String countryManufacturer = resultSet.getObject("m.country", String.class);
-        Manufacturer manufacturer = new Manufacturer(nameManufacturer, countryManufacturer);
-        manufacturer.setId(manufacturerId);
-        Car car = new Car(model, manufacturer);
+        Car car = new Car(model, getManufacturer(manufacturerId));
         car.setId(id);
         car.setDrivers(getDrivers(id));
         return car;
+    }
+
+    private Manufacturer getManufacturer(Long id) {
+        String getByIdQuery = "SELECT * FROM manufacturers"
+                + " WHERE id = ?"
+                + " AND deleted = false";
+        try (Connection connection = ConnectionUtil.getConnection();
+                 PreparedStatement preparedStatement = connection.prepareStatement(getByIdQuery)) {
+            preparedStatement.setLong(1, id);
+            ResultSet resultSet = preparedStatement.executeQuery();
+            Manufacturer manufacturer = null;
+            if (resultSet.next()) {
+                String nameManufacturer = resultSet.getObject("name", String.class);
+                String countryManufacturer = resultSet.getObject("country", String.class);
+                manufacturer = new Manufacturer(nameManufacturer, countryManufacturer);
+                manufacturer.setId(id);
+            }
+            return manufacturer;
+        } catch (SQLException e) {
+            throw new DataProcessingException("Cannot get manufacturer by id " + id, e);
+        }
     }
 
     private List<Driver> getDrivers(Long carId) {
